updated repositorys for better error handling

+++ Infrastructure/Repositories/CanteenWorkerRepository.cs
@@ -34,6 +34,11 @@ public class CanteenWorkerRepository : ICanteenWorkerRepository
 
     public async Task<CanteenWorker> UpdateAsync(CanteenWorker canteenWorker)
     {
+        var canteenWorkerToUpdate = await _context.CanteenWorkers.FirstOrDefaultAsync(x => x.Id == canteenWorker.Id);
+        if (canteenWorkerToUpdate == null)
+        {
+            throw new KeyNotFoundException($"CanteenWorker not found");
+        }
         _context.CanteenWorkers.Update(canteenWorker);
         await _context.SaveChangesAsync();
         return canteenWorker;
@@ -41,6 +46,12 @@ public class CanteenWorkerRepository : ICanteenWorkerRepository
 
     public async Task<CanteenWorker> DeleteAsync(CanteenWorker canteenWorker)
     {
+        var canteenWorkerToDelete = await _context.CanteenWorkers.FirstOrDefaultAsync(x => x.Id == canteenWorker.Id);
+        if (canteenWorkerToDelete == null)
+        {
+            throw new KeyNotFoundException("CanteenWorker not found");
+        }
+
         _context.CanteenWorkers.Remove(canteenWorker);
         await _context.SaveChangesAsync();
         return canteenWorker;
